refactor(Display): create display folder
